Move bracket subcomponents out into subfolder
